src/2_agenarchitec/requirements.txt
google-adk
google-genai
jupyter-server
